agents/codebuff/pyproject.toml
[project]
name = "codebuff"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "dagger-io",
    "ais-dagger-agents-config>=0.1.0,<0.2.0", 
    "pydantic-ai>=0.2.10",
    "pydantic[email]>=2.11.5",
    "pyyaml>=6.0.2",
    "simple-chalk>=0.1.0",
]


[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[tool.uv.sources]
dagger-io = { path = "sdk", editable = true }
